requirements.txt
anyio==3.6.1
bcrypt==3.2.2
black==22.6.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
ecdsa==0.18.0
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.2
watchfiles==0.16.1
websockets==10.3
